stock increased againing purchase

+++ src/main/java/com/inventory/inventoryservice/stock/model/StockDto.java
@@ -9,6 +9,6 @@ import lombok.*;
 public class StockDto {
 
     private Long productId;
-    private Integer currentStock;
+    private Double currentStock;
 
 }
